requirements.txt
opensearch-py
sentence-transformers
transformers
torch
python-dotenv
tqdm
numpy
matplotlib
scikit-learn
pyyaml
openai
